# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /copilot/proposals:
    get:
      operationId: listCopilotProposals
      tags: [copilot]
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
        - in: query
          name: shipmentId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of copilot proposals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CopilotProposal" }
    post:
      operationId: createCopilotProposal
      tags: [copilot]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CopilotProposalInput" }
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CopilotProposal" }
  /copilot/proposals/{id}:
    patch:
      operationId: updateCopilotProposal
      tags: [copilot]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CopilotProposalUpdate" }
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CopilotProposal" }
...
  /copilot/policies:
    get:
      operationId: listAutonomyPolicies
      tags: [copilot]
      responses:
        "200":
          description: List of autonomy policies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AutonomyPolicy" }
    put:
      operationId: upsertAutonomyPolicy
      tags: [copilot]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutonomyPolicyInput" }
      responses:
        "200":
          description: Upserted policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutonomyPolicy" }
...
components:
  schemas:
...
    CopilotProposal:
      type: object
      required: [id, shipmentId, triggerType, actionType, payload, reasoning, confidence, status, createdAt, updatedAt]
      properties:
        id:                  { type: integer }
        shipmentId:          { type: integer }
        triggerType:         { type: string }
        triggerRef:          { type: ["string", "null"] }
        actionType:          { type: string }
        payload:             { type: object, additionalProperties: true }
        reasoning:           { type: string }
        confidence:          { type: number }
        status:              { type: string }
        snoozedUntil:        { type: ["string", "null"], format: date-time }
        editedPayload:       { type: ["object", "null"], additionalProperties: true }
        userEditedContent:   { type: ["string", "null"], description: "Extracted edited draftBody text saved when user approves an edit" }
        editDistance:        { type: ["number", "null"], description: "Normalized 0–1 edit distance between AI draft and user edit; 0 = identical, 1 = fully rewritten" }
        auditTrail:          { type: array, items: { type: object, additionalProperties: true } }
        createdAt:           { type: string, format: date-time }
        updatedAt:           { type: string, format: date-time }
    CopilotProposalInput:
      type: object
      required: [shipmentId, triggerType, actionType, payload, reasoning]
      properties:
        shipmentId:  { type: integer }
        triggerType: { type: string }
        triggerRef:  { type: string }
        actionType:  { type: string }
        payload:     { type: object, additionalProperties: true }
        reasoning:   { type: string }
        confidence:  { type: number }
    CopilotProposalUpdate:
      type: object
      properties:
        status:        { type: string }
        editedPayload: { type: object, additionalProperties: true }
        snoozedUntil:  { type: string, format: date-time }
...
    AutonomyPolicy:
      type: object
      required: [id, policy, createdAt, updatedAt]
      properties:
        id:           { type: integer }
        supplierName: { type: ["string", "null"] }
        actionType:   { type: ["string", "null"] }
        policy:       { type: string }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
    AutonomyPolicyInput:
      type: object
      required: [policy]
      properties:
        supplierName: { type: string }
        actionType:   { type: string }
        policy:       { type: string }
...